some changes

+++ iterations.hpp
@@ -133,7 +133,7 @@ struct Iterations
     void prepareEdgeIndices();
 
     long get_index(uint i, uint j, uint k) const;
-    uint get_exact_index(uint i, uint j, uint k) const;
+    long get_exact_index(uint i, uint j, uint k) const;
 
     real x(uint i) const { return (i0 + i) * hx;}
     real y(uint j) const { return (j0 + j) * hy;}
